Add validation in mutations

+++ {{cookiecutter.project_slug}}/backend/{{cookiecutter.project_slug}}/users/mutations.py
@@ -6,15 +6,20 @@ from .models import User
 class UserMutation:
     @strawberry.mutation
     def updateUser(self, id: int, first_name: str, last_name: str ) -> UserType:
-        user = User.objects.get(id=id)
+        user = User.objects.filter(id=id).first()
         if user:
             user.first_name = first_name
             user.last_name = last_name
-        user.save
-        return user
+            user.save
+            return user
+        else:
+            return f"User Not found for id: {id}" 
 
     @strawberry.mutation
     def deleteUser(self, id: int) -> str:
-        user = User.objects.get(id=id)
-        user.delete()
-        return f"User with id {id} has been deleted."
+        user = User.objects.filter(id=id).first()
+        if user:
+            user.delete()
+            return f"User with id: {id} has been deleted."
+        else:
+            return f"User Not found for id: {id}"
